requirements.txt.txt

re
nltk
nltk.corpus.stopwords
nltk.tokenize.word_tokenize
sklearn.feature_extraction.text.TfidfVectorizer
sklearn.metrics.pairwise.cosine_similarity
pandas
IPython.display.Image
IPython.display.display
IPython.display.HTML
nltk.download('punkt')
numpy
pandas
matplotlib.pyplot
requests
sentence_transformers.SentenceTransformer
seaborn
sklearn.feature_extraction.text.CountVectorizer
sklearn.cluster.KMeans
sklearn.metrics.pairwise.cosine_similarity
sklearn.decomposition.PCA
streamlit
streamlit.components
PIL.Image
io.BytesIO
pickle
streamlit
PIL.Image
imghdr
os
